+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to MiniJinja are documented here.
   were always true, now they are only true if they are not empty.
 - Strings and safe strings no longer compare the same.
 - Changed list and map string output to have delimiters and debug printing.
+- Added `batch` and `slice` filter.
 
 # 0.10.0